Rename configDb to config-db
