+++ src/server/connectors/config-db/mockPathConfig.json
